requirements.txt
transformers == 4.35.2
torch == 2.1.1
nltk == 3.8.1
pandas == 2.1.3
langchain == 0.0.341
bitsandbytes == 0.41.2
openai == 0.27.7
scikit-learn == 1.3.2
python-dotenv == 1.0.0
sentence-transformers == 2.2.2
huggingface-hub == 0.19.4
#PyDataLoader == 0.1.3
#CT_METAL=1

#ctransformers --no-binary ctransformers